Enhance: Integrate NotificationModule into AccountModule and refactor AccountOwnerController for business onboarding functionality
