+++ src/modules/account/Account.module.ts
@@ -16,9 +16,10 @@ import { TagService } from '@/modules/account/app/impl/Tag.service';
 import { TagPublicController } from '@/modules/account/interfaces/Tag.public.controller';
 import { IBusinessService } from './app/IBusiness.service';
 import { BusinessService } from './app/impl/Business.service';
+import { NotificationModule } from '../notification/Notification.module';
 
 @Module({
-  imports: [AccountInfraModule, AuthModule],
+  imports: [AccountInfraModule, AuthModule, NotificationModule],
   controllers: [
     AccountUserController,
     AccountPublicController,
@@ -41,10 +42,10 @@ import { BusinessService } from './app/impl/Business.service';
       provide: IOnboardService,
       useClass: OnboardService,
     },
-    // {
-    //   provide: IBusinessService,
-    //   useClass: BusinessService,
-    // },
+    {
+      provide: IBusinessService,
+      useClass: BusinessService,
+    },
   ],
 })
 export class AccountModule {}
