+++ src/modules/post/Post.module.ts
@@ -6,7 +6,7 @@ import { AnalyticInfraModule } from '@/modules/analytic/infra/Analytic.infra.mod
 import { CommentController } from '@/modules/post/interfaces/Comment.controller';
 import { ICommentService } from '@/modules/post/app/IComment.service';
 import { CommentService } from '@/modules/post/app/impl/Comment.service';
-import { PostInfraModule } from '@/modules/post/infra/post.infra.module';
+import { PostInfraModule } from '@/modules/post/infra/Post.infra.module';
 
 @Module({
   imports: [PostInfraModule, AnalyticInfraModule],
